+++ R/Lesson.R
@@ -202,10 +202,14 @@ Lesson <- R6::R6Class("Lesson",
     #' frg <- Lesson$new(lesson_fragment())
     #' frg$validate_headings()
     validate_headings = function(verbose = TRUE) {
-      res <- purrr::map_dfr(self$episodes, 
-        ~.x$validate_headings(verbose = verbose, warn = FALSE),
-        .id = "episode"
+      res <- purrr::map(self$episodes, 
+        ~.x$validate_headings(verbose = verbose, warn = FALSE)
       )
+      if (requireNamespace("dplyr", quietly = TRUE)) {
+        res <- dplyr::bind_rows(res, .id = "episodes")
+      } else {
+        res <- do.call(rbind, res)
+      }
       throw_heading_warnings(res)
       invisible(res)
     },
@@ -233,7 +237,12 @@ Lesson <- R6::R6Class("Lesson",
     #' frg <- Lesson$new(lesson_fragment())
     #' frg$validate_links()
     validate_links = function() {
-      res <- purrr::map_dfr(self$episodes, ~.x$validate_links(warn = FALSE))
+      res <- purrr::map(self$episodes, ~.x$validate_links(warn = FALSE))
+      if (requireNamespace("dplyr", quietly = TRUE)) {
+        res <- dplyr::bind_rows(res)
+      } else {
+        res <- do.call(rbind, res)
+      }
       throw_link_warnings(res)
       invisible(res)
     }